Split Key module out of Cardano.Api.Typed
This is slightly different to the bits that had been split out before.
